python-gcloud/requirements.txt
google-cloud-resource-manager>=1.12.0
google-auth>=2.22.0
google-api-python-client>=2.100.0
rich>=13.0.0
